Fix capabilities tangle
Part of #142726
