+++ src/vs/workbench/contrib/terminal/browser/capabilities/commandDetectionCapability.ts
@@ -5,8 +5,7 @@
 
 import { Emitter } from 'vs/base/common/event';
 import { ILogService } from 'vs/platform/log/common/log';
-import { ICommandDetectionCapability, TerminalCapability } from 'vs/workbench/contrib/terminal/common/capabilities/capabilities';
-import { ITerminalCommand } from 'vs/workbench/contrib/terminal/common/terminal';
+import { ICommandDetectionCapability, TerminalCapability, ITerminalCommand } from 'vs/workbench/contrib/terminal/common/capabilities/capabilities';
 import { IBuffer, IMarker, Terminal } from 'xterm';
 
 interface ICurrentPartialCommand {
