+++ src/vs/workbench/contrib/terminal/common/capabilities/capabilities.ts
@@ -6,7 +6,6 @@
 import { Event } from 'vs/base/common/event';
 import { CwdDetectionCapability } from 'vs/workbench/contrib/terminal/common/capabilities/cwdDetectionCapability';
 import { NaiveCwdDetectionCapability } from 'vs/workbench/contrib/terminal/common/capabilities/naiveCwdDetectionCapability';
-import { ITerminalCommand, IXtermMarker } from 'vs/workbench/contrib/terminal/common/terminal';
 
 /**
  * Primarily driven by the shell integration feature, a terminal capability is the mechanism for
@@ -101,4 +100,26 @@ export interface IPartialCommandDetectionCapability {
 	readonly type: TerminalCapability.PartialCommandDetection;
 	readonly commands: readonly IXtermMarker[];
 	readonly onCommandFinished: Event<IXtermMarker>;
+}
+
+export interface ITerminalCommand {
+	command: string;
+	timestamp: number;
+	cwd?: string;
+	exitCode?: number;
+	marker?: IXtermMarker;
+	getOutput(): string | undefined;
+}
+
+/**
+ * A clone of the IMarker from xterm which cannot be imported from common
+ */
+export interface IXtermMarker {
+	readonly id: number;
+	readonly isDisposed: boolean;
+	readonly line: number;
+	dispose(): void;
+	onDispose: {
+		(listener: () => any): { dispose(): void };
+	};
 }
