+++ src/vs/workbench/contrib/terminal/common/terminal.ts
@@ -330,28 +330,6 @@ export interface IShellIntegration {
 	capabilities: ITerminalCapabilityStore;
 }
 
-export interface ITerminalCommand {
-	command: string;
-	timestamp: number;
-	cwd?: string;
-	exitCode?: number;
-	marker?: IXtermMarker;
-	getOutput(): string | undefined;
-}
-
-/**
- * A clone of the IMarker from xterm which cannot be imported from common
- */
-export interface IXtermMarker {
-	readonly id: number;
-	readonly isDisposed: boolean;
-	readonly line: number;
-	dispose(): void;
-	onDispose: {
-		(listener: () => any): { dispose(): void };
-	};
-}
-
 export interface INavigationMode {
 	exitNavigationMode(): void;
 	focusPreviousLine(): void;
