+++ src/vs/workbench/contrib/terminal/test/browser/capabilities/commandDetectionCapability.test.ts
@@ -8,7 +8,7 @@ import { timeout } from 'vs/base/common/async';
 import { Terminal } from 'xterm';
 import { CommandDetectionCapability } from 'vs/workbench/contrib/terminal/browser/capabilities/commandDetectionCapability';
 import { NullLogService } from 'vs/platform/log/common/log';
-import { ITerminalCommand } from 'vs/workbench/contrib/terminal/common/terminal';
+import { ITerminalCommand } from 'vs/workbench/contrib/terminal/common/capabilities/capabilities';
 
 async function writeP(terminal: Terminal, data: string): Promise<void> {
 	return new Promise<void>((resolve, reject) => {
